# taskify-front/assets/openapi.json
{
    "openapi": "3.1.0",
    "info": {
    "title": "Taskify Backend",
    "version": "0.1.0"
    },
    "paths": {
    "/api/v1/auth/forgot-password": {
    "post": {
    "tags": [
    "Auth"
    ],
    "summary": "Authcontroller.Forgot Password",
    "operationId": "AuthController_forgot_password_api_v1_auth_forgot_password_post",
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ForgotPassword"
    }
    }
    },
    "required": true
    },
    "responses": {
    "204": {
    "description": "Successful Response"
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/auth/logout": {
    "post": {
    "tags": [
    "Auth"
    ],
    "summary": "Authcontroller.Logout",
    "operationId": "AuthController_logout_api_v1_auth_logout_post",
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {}
    }
    }
    }
    },
    "security": [
    {
    "APIKeyCookie": []
    }
    ]
    }
    },
    "/api/v1/auth/login": {
    "post": {
    "tags": [
    "Auth"
    ],
    "summary": "Authcontroller.Password Login",
    "operationId": "AuthController_password_login_api_v1_auth_login_post",
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/LoginCredentials"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {}
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/auth/reset-password": {
    "post": {
    "tags": [
    "Auth"
    ],
    "summary": "Authcontroller.Reset Password",
    "operationId": "AuthController_reset_password_api_v1_auth_reset_password_post",
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ResetPasswordCredentials"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/UserRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    },
    "security": [
    {
    "APIKeyCookie": []
    }
    ]
    }
    },
    "/api/v1/auth/signup": {
    "post": {
    "tags": [
    "Auth"
    ],
    "summary": "Authcontroller.Signup",
    "operationId": "AuthController_signup_api_v1_auth_signup_post",
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/UserCreate"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/UserRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/auth/forgot-password-verify": {
    "post": {
    "tags": [
    "Auth"
    ],
    "summary": "Authcontroller.Verify Forgotten Password",
    "operationId": "AuthController_verify_forgotten_password_api_v1_auth_forgot_password_verify_post",
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/Body_AuthController_verify_forgotten_password_api_v1_auth_forgot_password_verify_post"
    }
    }
    },
    "required": true
    },
    "responses": {
    "204": {
    "description": "Successful Response"
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
...
    "/api/v1/boards/{board_id}/members/{member_id}": {
    "post": {
    "tags": [
    "Boards"
    ],
    "summary": "Boardcontroller.Add Board Member",
    "operationId": "BoardController_add_board_member_api_v1_boards__board_id__members__member_id__post",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "parameters": [
    {
    "name": "board_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Board Id"
    }
    },
    {
    "name": "member_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Member Id"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/BoardRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "delete": {
    "tags": [
    "Boards"
    ],
    "summary": "Boardcontroller.Remove Board Member",
    "operationId": "BoardController_remove_board_member_api_v1_boards__board_id__members__member_id__delete",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "parameters": [
    {
    "name": "board_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Board Id"
    }
    },
    {
    "name": "member_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Member Id"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/BoardRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/boards/": {
    "post": {
    "tags": [
    "Boards"
    ],
    "summary": "Boardcontroller.Create Board",
    "operationId": "BoardController_create_board_api_v1_boards__post",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/BoardCreate"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/BoardRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "get": {
    "tags": [
    "Boards"
    ],
    "summary": "Boardcontroller.Get User Boards",
    "operationId": "BoardController_get_user_boards_api_v1_boards__get",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "parameters": [
    {
    "name": "page",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "minimum": 1,
    "default": 1,
    "title": "Page"
    }
    },
    {
    "name": "size",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "maximum": 100,
    "minimum": 1,
    "default": 50,
    "title": "Size"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/Page_BoardRead_"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/boards/{board_id}": {
    "delete": {
    "tags": [
    "Boards"
    ],
    "summary": "Boardcontroller.Delete Board",
    "operationId": "BoardController_delete_board_api_v1_boards__board_id__delete",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "parameters": [
    {
    "name": "board_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Board Id"
    }
    }
    ],
    "responses": {
    "204": {
    "description": "Successful Response"
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "get": {
    "tags": [
    "Boards"
    ],
    "summary": "Boardcontroller.Get Board",
    "operationId": "BoardController_get_board_api_v1_boards__board_id__get",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "parameters": [
    {
    "name": "board_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Board Id"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/BoardRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "patch": {
    "tags": [
    "Boards"
    ],
    "summary": "Boardcontroller.Patch Board",
    "operationId": "BoardController_patch_board_api_v1_boards__board_id__patch",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "parameters": [
    {
    "name": "board_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Board Id"
    }
    }
    ],
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/BoardUpdate"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/BoardRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/boards/membered/all": {
    "get": {
    "tags": [
    "Boards"
    ],
    "summary": "Boardcontroller.Get A Membered Board",
    "operationId": "BoardController_get_a_membered_board_api_v1_boards_membered_all_get",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "parameters": [
    {
    "name": "page",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "minimum": 1,
    "default": 1,
    "title": "Page"
    }
    },
    {
    "name": "size",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "maximum": 100,
    "minimum": 1,
    "default": 50,
    "title": "Size"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/Page_BoardRead_"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/boards/public/{board_hash}": {
    "get": {
    "tags": [
    "Boards"
    ],
    "summary": "Boardcontroller.Get Public Board",
    "operationId": "BoardController_get_public_board_api_v1_boards_public__board_hash__get",
    "parameters": [
    {
    "name": "board_hash",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "title": "Board Hash"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/BoardRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/boards/{board_id}/change_public": {
    "patch": {
    "tags": [
    "Boards"
    ],
    "summary": "Boardcontroller.Update Publicity",
    "operationId": "BoardController_update_publicity_api_v1_boards__board_id__change_public_patch",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "parameters": [
    {
    "name": "board_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Board Id"
    }
    }
    ],
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/PublicBoardUpdate"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/BoardRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/lists/board/{board_id}": {
    "post": {
    "tags": [
    "Lists"
    ],
    "summary": "Listcontroller.Create List In Board",
    "operationId": "ListController_create_list_in_board_api_v1_lists_board__board_id__post",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "parameters": [
    {
    "name": "board_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Board Id"
    }
    }
    ],
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ListCreate"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ListRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "get": {
    "tags": [
    "Lists"
    ],
    "summary": "Listcontroller.Get Board Lists",
    "operationId": "ListController_get_board_lists_api_v1_lists_board__board_id__get",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "parameters": [
    {
    "name": "board_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Board Id"
    }
    },
    {
    "name": "page",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "minimum": 1,
    "default": 1,
    "title": "Page"
    }
    },
    {
    "name": "size",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "maximum": 100,
    "minimum": 1,
    "default": 50,
    "title": "Size"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/Page_ListRead_"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
...
    "/api/v1/lists/board/public/{board_hash}": {
    "get": {
    "tags": [
    "Lists"
    ],
    "summary": "Listcontroller.Get Public Lists",
    "operationId": "ListController_get_public_lists_api_v1_lists_board_public__board_hash__get",
    "parameters": [
    {
    "name": "board_hash",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "title": "Board Hash"
    }
    },
    {
    "name": "page",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "minimum": 1,
    "default": 1,
    "title": "Page"
    }
    },
    {
    "name": "size",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "maximum": 100,
    "minimum": 1,
    "default": 50,
    "title": "Size"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/Page_ListRead_"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
...
    },
    "components": {
    "schemas": {
    "BoardCreate": {
    "properties": {
    "title": {
    "type": "string",
    "title": "Title"
    },
    "description": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Description"
    },
    "background": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Background"
    },
    "owner_id": {
    "anyOf": [
    {
    "type": "string",
    "format": "uuid"
    },
    {
    "type": "null"
    }
    ],
    "title": "Owner Id"
    }
    },
    "type": "object",
    "required": [
    "title"
    ],
    "title": "BoardCreate"
    },
    "BoardMemeber": {
    "properties": {
    "id": {
    "type": "string",
    "format": "uuid",
    "title": "Id"
    },
    "created_at": {
    "type": "string",
    "format": "date-time",
    "title": "Created At"
    },
    "updated_at": {
    "type": "string",
    "format": "date-time",
    "title": "Updated At"
    },
    "email": {
    "type": "string",
    "title": "Email"
    },
    "username": {
    "type": "string",
    "title": "Username"
    },
    "full_name": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Full Name"
    },
    "avatar": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Avatar"
    }
    },
    "type": "object",
    "required": [
    "id",
    "created_at",
    "updated_at",
    "email",
    "username",
    "full_name",
    "avatar"
    ],
    "title": "BoardMemeber"
    },
    "BoardRead": {
    "properties": {
    "id": {
    "type": "string",
    "format": "uuid",
    "title": "Id"
    },
    "created_at": {
    "type": "string",
    "format": "date-time",
    "title": "Created At"
    },
    "updated_at": {
    "type": "string",
    "format": "date-time",
    "title": "Updated At"
    },
    "title": {
    "type": "string",
    "title": "Title"
    },
    "description": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Description"
    },
    "background": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Background"
    },
    "owner_id": {
    "type": "string",
    "format": "uuid",
    "title": "Owner Id"
    },
    "is_public": {
    "type": "boolean",
    "title": "Is Public",
    "default": false
    },
    "public_hash": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Public Hash"
    },
    "members": {
    "items": {
    "$ref": "#/components/schemas/BoardMemeber"
    },
    "type": "array",
    "title": "Members"
    },
    "owner": {
    "$ref": "#/components/schemas/BoardMemeber"
    }
    },
    "type": "object",
    "required": [
    "id",
    "created_at",
    "updated_at",
    "title",
    "description",
    "background",
    "owner_id",
    "members",
    "owner"
    ],
    "title": "BoardRead"
    },
    "BoardUpdate": {
    "properties": {
    "title": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Title"
    },
    "description": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Description"
    },
    "background": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Background"
    }
    },
    "type": "object",
    "title": "BoardUpdate"
    },
    "Body_AuthController_verify_forgotten_password_api_v1_auth_forgot_password_verify_post": {
    "properties": {
    "new_password": {
    "type": "string",
    "title": "New Password"
    },
    "token": {
    "type": "string",
    "title": "Token"
    }
    },
    "type": "object",
    "required": [
    "new_password",
    "token"
    ],
    "title": "Body_AuthController_verify_forgotten_password_api_v1_auth_forgot_password_verify_post"
    },
...
    "ForgotPassword": {
    "properties": {
    "email": {
    "type": "string",
    "title": "Email"
    }
    },
    "type": "object",
    "required": [
    "email"
    ],
    "title": "ForgotPassword"
    },
    "HTTPValidationError": {
    "properties": {
    "detail": {
    "items": {
    "$ref": "#/components/schemas/ValidationError"
    },
    "type": "array",
    "title": "Detail"
    }
    },
    "type": "object",
    "title": "HTTPValidationError"
    },
    "ListCreate": {
    "properties": {
    "title": {
    "type": "string",
    "title": "Title"
    },
    "position": {
    "anyOf": [
    {
    "type": "integer"
    },
    {
    "type": "null"
    }
    ],
    "title": "Position"
    }
    },
    "type": "object",
    "required": [
    "title"
    ],
    "title": "ListCreate"
    },
    "ListRead": {
    "properties": {
    "id": {
    "type": "string",
    "format": "uuid",
    "title": "Id"
    },
    "created_at": {
    "type": "string",
    "format": "date-time",
    "title": "Created At"
    },
    "updated_at": {
    "type": "string",
    "format": "date-time",
    "title": "Updated At"
    },
    "title": {
    "type": "string",
    "title": "Title"
    },
    "position": {
    "type": "integer",
    "title": "Position"
    },
    "board_id": {
    "type": "string",
    "format": "uuid",
    "title": "Board Id"
    }
    },
    "type": "object",
    "required": [
    "id",
    "created_at",
    "updated_at",
    "title",
    "position",
    "board_id"
    ],
    "title": "ListRead"
    },
...
    "LoginCredentials": {
    "properties": {
    "email": {
    "type": "string",
    "title": "Email"
    },
    "password": {
    "type": "string",
    "title": "Password"
    }
    },
    "type": "object",
    "required": [
    "email",
    "password"
    ],
    "title": "LoginCredentials"
    },
...
    "OAuthRead": {
    "properties": {
    "id": {
    "type": "string",
    "format": "uuid",
    "title": "Id"
    },
    "user_id": {
    "type": "string",
    "format": "uuid",
    "title": "User Id"
    },
    "oauth_name": {
    "type": "string",
    "title": "Oauth Name"
    },
    "account_id": {
    "type": "string",
    "title": "Account Id"
    },
    "account_email": {
    "type": "string",
    "title": "Account Email"
    }
    },
    "type": "object",
    "required": [
    "id",
    "user_id",
    "oauth_name",
    "account_id",
    "account_email"
    ],
    "title": "OAuthRead"
    },
...
    "Page_BoardRead_": {
    "properties": {
    "items": {
    "items": {
    "$ref": "#/components/schemas/BoardRead"
    },
    "type": "array",
    "title": "Items"
    },
    "total": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 0
    },
    {
    "type": "null"
    }
    ],
    "title": "Total"
    },
    "page": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 1
    },
    {
    "type": "null"
    }
    ],
    "title": "Page"
    },
    "size": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 1
    },
    {
    "type": "null"
    }
    ],
    "title": "Size"
    },
    "pages": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 0
    },
    {
    "type": "null"
    }
    ],
    "title": "Pages"
    }
    },
    "type": "object",
    "required": [
    "items",
    "page",
    "size"
    ],
    "title": "Page[BoardRead]"
    },
...
    "Page_ListRead_": {
    "properties": {
    "items": {
    "items": {
    "$ref": "#/components/schemas/ListRead"
    },
    "type": "array",
    "title": "Items"
    },
    "total": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 0
    },
    {
    "type": "null"
    }
    ],
    "title": "Total"
    },
    "page": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 1
    },
    {
    "type": "null"
    }
    ],
    "title": "Page"
    },
    "size": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 1
    },
    {
    "type": "null"
    }
    ],
    "title": "Size"
    },
    "pages": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 0
    },
    {
    "type": "null"
    }
    ],
    "title": "Pages"
    }
    },
    "type": "object",
    "required": [
    "items",
    "page",
    "size"
    ],
    "title": "Page[ListRead]"
    },
    "PublicBoardUpdate": {
    "properties": {
    "is_public": {
    "type": "boolean",
    "title": "Is Public"
    }
    },
    "type": "object",
    "required": [
    "is_public"
    ],
    "title": "PublicBoardUpdate"
    },
    "ResetPasswordCredentials": {
    "properties": {
    "old_password": {
    "type": "string",
    "title": "Old Password"
    },
    "new_password": {
    "type": "string",
    "title": "New Password"
    }
    },
    "type": "object",
    "required": [
    "old_password",
    "new_password"
    ],
    "title": "ResetPasswordCredentials"
    },
    "UserCreate": {
    "properties": {
    "email": {
    "type": "string",
    "title": "Email"
    },
    "username": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Username"
    },
    "full_name": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Full Name"
    },
    "password": {
    "type": "string",
    "title": "Password"
    },
    "role_id": {
    "anyOf": [
    {
    "type": "integer"
    },
    {
    "type": "null"
    }
    ],
    "title": "Role Id"
    }
    },
    "type": "object",
    "required": [
    "email",
    "password"
    ],
    "title": "UserCreate"
    },
    "UserRead": {
    "properties": {
    "id": {
    "type": "string",
    "format": "uuid",
    "title": "Id"
    },
    "created_at": {
    "type": "string",
    "format": "date-time",
    "title": "Created At"
    },
    "updated_at": {
    "type": "string",
    "format": "date-time",
    "title": "Updated At"
    },
    "email": {
    "type": "string",
    "title": "Email"
    },
    "username": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Username"
    },
    "full_name": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Full Name"
    },
    "avatar": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Avatar"
    },
    "role_id": {
    "type": "integer",
    "title": "Role Id"
    },
    "is_active": {
    "type": "boolean",
    "title": "Is Active"
    },
    "oauth_accounts": {
    "items": {
    "$ref": "#/components/schemas/OAuthRead"
    },
    "type": "array",
    "title": "Oauth Accounts",
    "default": []
    }
    },
    "type": "object",
    "required": [
    "id",
    "created_at",
    "updated_at",
    "email",
    "username",
    "full_name",
    "avatar",
    "role_id",
    "is_active"
    ],
    "title": "UserRead"
    },
...
    "ValidationError": {
    "properties": {
    "loc": {
    "items": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "integer"
    }
    ]
    },
    "type": "array",
    "title": "Location"
    },
    "msg": {
    "type": "string",
    "title": "Message"
    },
    "type": {
    "type": "string",
    "title": "Error Type"
    }
    },
    "type": "object",
    "required": [
    "loc",
    "msg",
    "type"
    ],
    "title": "ValidationError"
    }
    },
    "securitySchemes": {
    "APIKeyCookie": {
    "type": "apiKey",
    "in": "cookie",
    "name": "taskify_auth"
    }
    }
    }
    }